import + export + dashboard Emplye et Responsable

+++ Back_end/app/services/delete_stock_service.py
@@ -4,20 +4,14 @@ from app.models.Stock import Stock
 
 
 def delete_stock(stock_id: int):
-
-    # 1. Rechercher l'objet en base de données
     db_stock = Stock.query.filter(Stock.id == stock_id).first()
+    if not db_stock:
+        abort(404, description=f"Le stock avec l'id {stock_id} n'existe pas.")
 
-    # 2. Vérifier s'il existe
-    # if not db_stock:
-    #     abort(404, description=f"Le stock avec l'id {stock_id} n'existe pas.")
-
-    # 3. Supprimer l'objet
-    db.session.delete(db_stock)
-
-    # 4. Valider la transaction
     try:
+        db.session.delete(db_stock)
         db.session.commit()
+        return db_stock
     except Exception as e:
         db.session.rollback()
         print(f"Erreur : {e}")
